test: address report coverage review
